1.调整了E2的操作前SCL位输出不为低电平的BUG 2.修改了DATA_FLASH写入的未擦除的BUG 3.增加了生成MSG数据格式的函数 4.调整XML文件头的记录条数根据数据记录条数变化 5.修改了存储的数据都为第一通道的BUG 6.增加了获取记录条数的函数

+++ User/IEC101.h
@@ -74,7 +74,7 @@ struct IEC101_STRUCT
   u32 wRecvLen;
   u8 PReMsgType;
   u8 byReason;
-  u8 byPSGenStep;
+  u32 byPSGenStep;
   u8 frameno;
   u8 TypeProtocol;
   u8 byPSDdStep;
